requirements.txt
requests
sseclient-py
pystray
Pillow
keyring
pyinstaller
